refactor: rename package resource in base, that is equal to the class name
we don't use a resource type called "resource" but a base class called
"Base".
