+++ src/adhocracy_core/adhocracy_core/resources/itemversion.py
@@ -5,7 +5,7 @@ from adhocracy_core.interfaces import IItemVersion
 from adhocracy_core.interfaces import SheetToSheet
 from adhocracy_core.resources import add_resource_type_to_registry
 from adhocracy_core.resources import resource_meta
-from adhocracy_core.resources.resource import Base
+from adhocracy_core.resources.base import Base
 from adhocracy_core.sheets.versions import IVersionable
 from adhocracy_core.utils import get_sheet
 from adhocracy_core.utils import find_graph
