+++ src/adhocracy_core/adhocracy_core/resources/pool.py
@@ -11,8 +11,8 @@ import adhocracy_core.sheets.metadata
 from adhocracy_core.interfaces import IPool
 from adhocracy_core.resources import add_resource_type_to_registry
 from adhocracy_core.resources import resource_meta
-from adhocracy_core.resources.resource import Base
-from adhocracy_core.resources.resource import Length
+from adhocracy_core.resources.base import Base
+from adhocracy_core.resources.base import Length
 
 
 class IBasicPool(IPool):
