+++ src/adhocracy_core/adhocracy_core/resources/simple.py
@@ -4,7 +4,7 @@ import adhocracy_core.sheets.name
 import adhocracy_core.sheets.metadata
 from adhocracy_core.resources import add_resource_type_to_registry
 from adhocracy_core.resources import resource_meta
-from adhocracy_core.resources.resource import Base
+from adhocracy_core.resources.base import Base
 
 
 simple_meta = resource_meta._replace(
