+++ src/adhocracy_core/adhocracy_core/resources/tag.py
@@ -2,7 +2,7 @@
 from adhocracy_core.interfaces import ITag
 from adhocracy_core.resources import add_resource_type_to_registry
 from adhocracy_core.resources import resource_meta
-from adhocracy_core.resources.resource import Base
+from adhocracy_core.resources.base import Base
 import adhocracy_core.sheets.name
 import adhocracy_core.sheets.tags
 import adhocracy_core.sheets.metadata
